Back to component initial names to try to preserve Git history
